Extract multiple types out of Cluster.
The goal of this refactoring was to split apart the ball of mud that was
the Cluster type. This new version has two types, NewCluster and
ExistingCluster, each with only the fields needed for their specific
tasks, and common fields extracted into a new, slimmer Cluster type. The
various file paths that were originally calculated in each type's
constructor is now done lazily through a method. Because each type only
gets fields it needs, many Option<T> fields are now simply T.