Format EBS volume from a script.
Using a negative ConditionPathExists to determine whether or not the EBS
volume should be formatted didn't work, because in the case where the
path existed, it would cause the whole systemd unit to return an error
code, which in turn would prevent depending systemd units to start.

With this script, exit code 0 will be returned even if the volume has
been formatted.